+++ lib/browser/base.rb
@@ -213,6 +213,12 @@ module Browser
       Maxthon.new(ua).match? && detect_version?(full_version, expected_version)
     end
 
+    # Detect if browser is Sougou
+    def sougou_browser?(expected_version = nil)
+      SougouBrowser.new(ua).match? &&
+        detect_version?(full_version, expected_version)
+    end
+
     # Detect if browser is Google Search App
     def google_search_app?(expected_version = nil)
       ua =~ /GSA/ && detect_version?(full_version, expected_version)